requirements.txt
Django==4.2.14
PyMySQL==1.1.0
python-dotenv==1.0.1
requests==2.32.3
web3==6.19.0
